requirements.txt
opentelemetry-sdk
opentelemetry-exporter-otlp

